+++ models.py
@@ -867,7 +867,7 @@ class BratsNewSegmentationNet(nn.Module):
             pred_y = self(x.to(self.device))
             if train:
                 batch_loss = F.cross_entropy(
-                    pred_y, torch.squeeze(y.to(self.device))
+                    pred_y, torch.squeeze(y.type(torch.long).to(self.device))
                 )
                 batch_loss.backward()
                 self.optimizer_alg.step()
